Finalize `trackStore` implementation
